+++ cut_cross_entropy/indexed_dot.py
@@ -43,8 +43,8 @@ def _indexed_neg_dot_forward_kernel(
     group_id = pid // num_d_in_group
     first_pid_b = group_id * GROUP_B
     group_size_b = min(num_b_chunks - first_pid_b, GROUP_B)
-    pid_b = first_pid_b + ((pid % num_d_in_group) % group_size_b)
-    pid_d = (pid % num_d_in_group) // group_size_b
+    pid_b = (first_pid_b + ((pid % num_d_in_group) % group_size_b)).to(tl.int64)
+    pid_d = ((pid % num_d_in_group) // group_size_b).to(tl.int64)
 
     offs_b = (tl.arange(0, BLOCK_B) + pid_b * BLOCK_B).to(tl.int64)
     if HAS_VALIDS:
